# tests/Formbase.Host.Tests/Contracts/served-openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Formbase.Host | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "{origin}"
    }
  ],
  "paths": {
    "/formtypes/{type}/documents": {
      "post": {
        "tags": [
          "DocumentEndpoints"
        ],
        "summary": "Accepts a document into the raw store",
        "description": "The document is stored verbatim; no declaration is required and none is consulted. An object naming a property twice is refused with 400. Send an Idempotency-Key header to make re-submission safe: the same key with the same request returns the same document without appending a second copy. A key already used for another request (another form type, record key, or body) is refused with 422. Name a \u0060recordKey\u0060 to make the document a correction of that record: the projection shows only the record\u0027s latest document. Without one, the document is a record of its own.",
        "operationId": "AcceptDocument",
        "parameters": [
          {
            "name": "type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "recordKey",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Formbase-Namespace",
            "in": "header",
            "description": "Which namespace the request is addressed to. Omit it to address the host\u0027s own; naming one this host does not serve is refused rather than answered from the one it does.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcceptedDocumentResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "DocumentEndpoints"
        ],
        "summary": "Reads a page of a form type\u0027s raw stream",
        "description": "Documents come back as they were accepted, oldest first, after the \u0060after\u0060 watermark (default 0, the start of the stream) \u2014 whether or not the form type has a declaration or a projection, so fields nothing has declared are readable here. \u0060limit\u0060 defaults to 100 and is at most 1000; \u0060limit=0\u0060 reads only \u0060rawHead\u0060. The page never reaches past \u0060rawHead\u0060, so a caller continues from the last watermark it received until that watermark equals \u0060rawHead\u0060.",
        "operationId": "ListDocuments",
        "parameters": [
          {
            "name": "type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "after",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int64"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          },
          {
            "name": "Formbase-Namespace",
            "in": "header",
            "description": "Which namespace the request is addressed to. Omit it to address the host\u0027s own; naming one this host does not serve is refused rather than answered from the one it does.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentPageResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/documents/{id}": {
      "get": {
        "tags": [
          "DocumentEndpoints"
        ],
        "summary": "Reads a stored document",
        "description": "Reads from the raw store, which is the source of truth. This answers whether or not the form type has ever been projected.",
        "operationId": "GetDocument",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Formbase-Namespace",
            "in": "header",
            "description": "Which namespace the request is addressed to. Omit it to address the host\u0027s own; naming one this host does not serve is refused rather than answered from the one it does.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoredDocumentResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AcceptedDocumentResponse": {
        "required": [
          "documentId",
          "formType"
        ],
        "type": "object",
        "properties": {
          "documentId": {
            "type": "string",
            "format": "uuid"
          },
          "formType": {
            "type": "string"
          },
          "recordKey": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "DocumentPageResponse": {
        "required": [
          "documents",
          "rawHead"
        ],
        "type": "object",
        "properties": {
          "documents": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StoredDocumentResponse"
            }
          },
          "rawHead": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          }
        }
      },
      "JsonElement": {},
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "StoredDocumentResponse": {
        "required": [
          "documentId",
          "formType",
          "watermark",
          "appendedAt",
          "body",
          "recordKey",
          "retired"
        ],
        "type": "object",
        "properties": {
          "documentId": {
            "type": "string",
            "format": "uuid"
          },
          "formType": {
            "type": "string"
          },
          "watermark": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "appendedAt": {
            "type": "string",
            "format": "date-time"
          },
          "body": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/JsonElement"
              }
            ]
          },
          "recordKey": {
            "type": [
              "null",
              "string"
            ]
          },
          "retired": {
            "type": "boolean"
          }
        }
      }
    }
  },
...
}